Day 18 B
Part 2: use a formula for the LARGE polygon instead of scanline fill.

Shoelace etc is overkill and would require to compute coordinates.
Trapezoid formula is immediate as we already have the edges.

The difficulty is into combining the edge (trench) and the interior when
we only get the edge length of the "middle" of the trench.

We process the original path to construct a new path corresponding
to its exterior edge and compute the area on this one.